% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{np_lt_residual}
\alias{np_lt_residual}
\title{Nonparametric Estimator for the Univariate Laplace Transform of Residual Life}
\usage{
np_lt_residual(x, s, t = 0)
}
\arguments{
\item{x}{Numeric vector; observed sample.}

\item{s}{Non-negative Laplace parameter.}

\item{t}{Truncation time (default 0).}
}
\value{
Scalar numeric estimate of \eqn{L_X(s,t)}.
}
\description{
Given a sample \eqn{X_1, \ldots, X_n}, estimates the Laplace transform of
the residual life using the empirical survival function:
\deqn{\hat{L}_X(s,t) =
  \frac{\displaystyle\sum_{i:\,X_i > t} e^{-s X_i}}
       {\#\{i : X_i > t\}}.}
}
\examples{
set.seed(1)
x <- rexp(300, rate = 1)

# Estimate at s=1, t=0: true value 1/(1+1) = 0.5
np_lt_residual(x, s = 1, t = 0)

# Estimate at s=1, t=0.5: true value still approx 0.5 (memoryless)
np_lt_residual(x, s = 1, t = 0.5)

}
\seealso{
\code{\link{lt_residual}}
}
