% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_blt_reversed}
\alias{plot_blt_reversed}
\title{Plot Bivariate Laplace Transform of Reversed Residual Lives}
\usage{
plot_blt_reversed(
  cdf_fn,
  cdf_fn2 = NULL,
  s1 = 1,
  t2 = 0.5,
  t1_grid = seq(0.1, 0.9, by = 0.05),
  theta = 0,
  xlab = expression(t[1]),
  ylab = expression(L[t[1] * "|" * t[2]](s[1])),
  main = "Bivariate LT of Reversed Residual Lives",
  col1 = "darkgreen",
  col2 = "darkorange",
  lwd = 2,
  legend_labels = c("Distribution 1", "Distribution 2")
)
}
\arguments{
\item{cdf_fn}{Joint CDF function.}

\item{cdf_fn2}{Optional second CDF for comparison.}

\item{s1}{Laplace parameter.}

\item{t2}{Fixed second truncation time.}

\item{t1_grid}{Grid of first truncation times.}

\item{theta}{FGM parameter (used if \code{cdf_fn = NULL}).}

\item{xlab, ylab, main}{Plot labels.}

\item{col1, col2}{Line colours.}

\item{lwd}{Line width.}

\item{legend_labels}{Legend labels.}
}
\value{
Invisibly returns the data frame used for plotting.
}
\description{
Plots the reversed-life Laplace transform \eqn{L_{t_1|t_2}(s_1)} as a
function of \eqn{t_1} for fixed \eqn{s_1} and \eqn{t_2}.
}
\examples{
cX <- function(x1, x2) pfgm_biv(x1, x2, theta = 0.2)
cY <- function(x1, x2) pfgm_biv(x1, x2, theta = 0.7)
plot_blt_reversed(cX, cY, s1 = 1, t2 = 0.5,
                  legend_labels = c("theta=0.2", "theta=0.7"))

}
